:pig_nose: 评论功能的添加,token修改完善，对已观看过的视频不再进行记录，以及个人信息的存储

+++ src/app/httpInterceptor/http-inter.service.ts
@@ -4,12 +4,14 @@ import {Injectable} from '@angular/core';
 import {tap} from 'rxjs/operators';
 import {environment} from '../../environments/environment';
 import {Router} from '@angular/router';
+import {CookieService} from 'ngx-cookie-service';
 
 @Injectable()
 export class HttpInter implements HttpInterceptor {
   private baseUrl: string;
+  private token: string;
 
-  constructor(private router: Router) {
+  constructor(private router: Router, private cookieService: CookieService) {
     if (environment.production) {
       this.baseUrl = 'http://106.12.195.114:8085';
     } else {
@@ -26,6 +28,12 @@ export class HttpInter implements HttpInterceptor {
     let request: HttpRequest<any>;
     // 登录无需要token
     console.log('URL ', req.url);
+    if (this.cookieService.get('token')) {
+      this.token = this.cookieService.get('token');
+    } else {
+      this.token = sessionStorage.getItem('token');
+    }
+    console.log('Token is ', this.token);
     if (req.url === '/user/login' || req.url === '/report/down') {
       request = req.clone({
         url: `${this.baseUrl}${req.url}`,
@@ -36,7 +44,7 @@ export class HttpInter implements HttpInterceptor {
         url: `${this.baseUrl}${req.url}`,
         headers: req.headers.set(
           'Authorization',
-          localStorage.getItem('token'),
+          this.token,
         ),
       });
       console.log('Token ', localStorage.getItem('token'));
@@ -45,13 +53,12 @@ export class HttpInter implements HttpInterceptor {
       tap(
         event => (ok = event instanceof HttpErrorResponse ? 'success' : ''),
         (error: HttpErrorResponse) => {
-          switch (error.status) {
-            case 403:
-              // 登录错误，返回
-              console.log('403了');
-              localStorage.clear();
-              this.router.navigateByUrl('/login');
-              break;
+          if (error.status === 403) {
+            console.log('403了');
+            localStorage.clear();
+            sessionStorage.clear();
+            this.cookieService.deleteAll();
+            this.router.navigateByUrl('/login');
           }
           console.log('拦截器的error', error);
         },
